day8/project_day_8/requirements.txt
pydantic==2.*
httpx
python-dotenv
litellm
flask
flask-cors
requests
pytest
